udp: move newly generic code so it can be reused
